Add tests and move to detecting note transitions as boundaries
Previously I was peak picking from the change_signal. This commit
changes things so that the width of the peak in the change_signal is
considered to be a transition and is therefore ignored. This seems to
help with cleaning up the median note measurements as the slides
tended to skew these. The onset timing might be slightly less accurate
and long slides into notes may be ignored. Also the defaults were
adjusted in favour of increased sensitivity as the new method seems to
be more robust at selecting and merging the correct notes.